backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx>=0.27.0
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
redis[hiredis]==5.1.0
miniopy-async==1.21.1
python-multipart==0.0.9
# WP5: protobuf decoding for C++ UDP publisher
protobuf==5.29.4
grpcio-tools==1.71.0
